Complete rewrite to use sources sigh, REFN is a failure. only 10 chars to work with without crashing PAF
thought about adding a note, and could...
thought about requiring user to specify REFN, and could...

I guess the best thing though is to use a SOURce record. sources can be cited for
FAM and INDI records at the record level. possible con is merging will merge in
that source even if you change everything about that original record you
imported (i.e. if it's a name-only entry and the name is spelled wrong). still,
this is not a unique problem and it is shared by the NOTE and REFN approaches.
A nice thing about the source is if you want you can delete the source citation
or even the entire source (and all references to it) as a unit.

So, this means I needed to come up with a source citation that is highly unlikely
to collide. tag uri comes to mind, but : and @ aren't legal in an xref id.
Maybe s/[:@]/_/, but then it's not really a tag uri anymore. Perhaps a hash of
the gedcom. Yeah, that sounds great, an MD5 hash. Too bad PAF doesn't like long
xref IDs. (I couldn't find a number of characters in the GEDCOM spec, so I just
used "GEDTAG" which is unlikely to collide with anything but itself - maybe I
should print a warning in case of collision, and probably I should provide an
option to override it)

0 @GEDTAG@ SOUR
 1 AUTH <submitter name>
 1 TITL <gedcom filename>
 1 NOTE Blah blah gedtag (http://hans.fugal.net/src/gedtag)

0 {INDI,FAM} ...
 1 SOUR @GEDTAG@
  2 DATA
   3 DATE <gedcom date, if exists>
  2 PAGE <xref>